複数テーブルに保存実装

+++ db/migrate/20201113091125_create_addresses.rb
@@ -5,7 +5,7 @@ class CreateAddresses < ActiveRecord::Migration[6.0]
       t.integer :region_id, null: false
       t.string :city, null: false
       t.string :house, null: false
-      t.integer :telephone, null: false
+      t.bigint :telephone, null: false
       t.references :record, foreign_key: true
       t.timestamps
     end
